+++ src/Components/Component/Header.jsx
@@ -7,41 +7,11 @@ import {
 } from "@material-tailwind/react";
 import { Bars3Icon, XMarkIcon } from "@heroicons/react/16/solid";
 import { Link } from "react-router-dom";
+import UserProfile from "./UserProfile";
 
-import {
-  Avatar,
-  Button,
-  Menu,
-  MenuHandler,
-  MenuItem,
-  MenuList,
-} from "@material-tailwind/react";
-import {
-  Cog6ToothIcon,
-  PowerIcon,
-  UserCircleIcon,
-} from "@heroicons/react/24/solid";
 
-// profile menu component
-const profileMenuItems = [
-  {
-    label: "My Profile",
-    icon: UserCircleIcon,
-  },
-  {
-    label: "Edit Profile",
-    icon: Cog6ToothIcon,
-  },
-
-  {
-    label: "Sign Out",
-    icon: PowerIcon,
-  },
-];
 const NavList = () => {
-  const [isMenuOpen, setIsMenuOpen] = React.useState(false);
 
-  const closeMenu = () => setIsMenuOpen(false);
   return (
     <ul className="my-2 flex flex-col gap-2 lg:mb-0 lg:mt-0 lg:flex-row lg:items-center lg:gap-6">
       <Typography
@@ -96,54 +66,7 @@ const NavList = () => {
           Sign Up
         </Link>
       </Typography>
-      <Menu open={isMenuOpen} handler={setIsMenuOpen} placement="bottom-end">
-        <MenuHandler>
-          <Button
-            variant="text"
-            color="blue-gray"
-            className="flex items-center rounded-full p-0"
-          >
-            <Avatar
-              variant="circular"
-              size="md"
-              alt="tania andrew"
-              withBorder={true}
-              color="blue-gray"
-              className=" p-0.5"
-              src="https://docs.material-tailwind.com/img/face-2.jpg"
-            />
-          </Button>
-        </MenuHandler>
-        <MenuList className="p-1">
-          {profileMenuItems.map(({ label, icon }, key) => {
-            const isLastItem = key === profileMenuItems.length - 1;
-            return (
-              <MenuItem
-                key={label}
-                onClick={closeMenu}
-                className={`flex items-center gap-2 rounded ${
-                  isLastItem
-                    ? "hover:bg-red-500/10 focus:bg-red-500/10 active:bg-red-500/10"
-                    : ""
-                }`}
-              >
-                {React.createElement(icon, {
-                  className: `h-4 w-4 ${isLastItem ? "text-red-500" : ""}`,
-                  strokeWidth: 2,
-                })}
-                <Typography
-                  as="span"
-                  variant="small"
-                  className="font-normal"
-                  color={isLastItem ? "red" : "inherit"}
-                >
-                  {label}
-                </Typography>
-              </MenuItem>
-            );
-          })}
-        </MenuList>
-      </Menu>
+     <UserProfile/>
     </ul>
   );
 };
