+++ src/Components/Pages/User/Login.jsx
@@ -2,28 +2,29 @@ import React, { useState } from "react";
 import { Card, Input, Button, Typography } from "@material-tailwind/react";
 import axios from "axios";
 import { useNavigate } from "react-router-dom";
-const Login = ({ allUsers }) => {
+const Login = ({ allUsers },  {setIsLoged },  isLoged ) => {
   const [user, setUser] = useState({
-    userName: "",
+    email: "",
     password: "",
   });
-  const [checkUserName, setCheckUserName] = useState(false);
+  const navigate = useNavigate();
+  const [checkEmail, setCheckEmail] = useState(false);
   const [checkPassword, setCheckPassword] = useState(false);
   const reset = () => {
-    setCheckUserName(false);
+    setCheckEmail(false);
     setCheckPassword(false);
   };
   const handelForm = (e) => {
     e.preventDefault();
-    if (user.userName == "") {
+    if (user.email == "") {
       reset();
-      setCheckUserName(true);
+      setCheckEmail(true);
     } else if (user.password == "") {
       reset();
       setCheckPassword(true);
     } else {
       const userInfo = {
-        name: user.userName,
+        email: user.email,
         password: user.password,
         gender: "",
       };
@@ -31,22 +32,28 @@ const Login = ({ allUsers }) => {
         method: "post",
         url: import.meta.env.VITE_Users,
         data: userInfo,
+      }).then(() => {
+        setIsLoged(true);
+        localStorage.id = user.id;
+        navigate("/");
       });
+      
+
+      //   const loginUser = allUsers.find(
+      //   ({ email, password }) => user.email == email && user.password == password
+      // );
+      // if (loginUser) {
+      //   localStorage.ok = loginUser.id;
+      //   useNavigate("/");
+      // }
     }
-    // const loginUser = allUsers.find(
-    //   ({ name, password }) => user.name == name && user.password == password
-    // );
-    // if (loginUser) {
-    //   localStorage.ok = loginUser.id;
-    //   useNavigate("/");
-    // }
   };
   return (
     <div className="flex flex-col justify-center items-center px-[4em] gap-[4em]">
       <div>
         <Card color="transparent" shadow={false}>
           <Typography variant="h4" color="blue-gray">
-            Login{" "}
+            Login
           </Typography>
           <form
             onSubmit={(e) => handelForm(e)}
@@ -54,15 +61,10 @@ const Login = ({ allUsers }) => {
           >
             <div className="mb-1 flex flex-col gap-6 justify-center">
               <Input
-                error={checkUserName}
-                label="User Name"
-                value={user.userName}
-                onChange={(e) =>
-                  setUser({
-                    ...user,
-                    userName: e.target.value,
-                  })
-                }
+                error={checkEmail}
+                label="Email"
+                value={user.email}
+                onChange={(e) => setUser({ ...user, email: e.target.value })}
               />
               <Input
                 error={checkPassword}
@@ -70,12 +72,7 @@ const Login = ({ allUsers }) => {
                 type="password"
                 placeholder="********"
                 value={user.password}
-                onChange={(e) =>
-                  setUser({
-                    ...user,
-                    password: e.target.value,
-                  })
-                }
+                onChange={(e) => setUser({ ...user, password: e.target.value })}
               />
             </div>
             <Button color="green" type="submit" className="mt-6" fullWidth>
