--- src/Components/Pages/User/signin.jsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const Signin = () => {
-  return (
-    <div>
-      <h1>Signin</h1>
-    </div>
-  )
-}
-
-export default Signin
